Added more losses, created new submodule to contain new losses
