+++ smartpy/losses/__init__.py
@@ -1,0 +1,3 @@
+from .classification_losses import *
+from .distribution_losses import *
+from .reconstruction_losses import *
